pyproject.toml
[tool.poetry]
name = "connect-hub"
version = "0.1.0"
description = ""
authors = ["Eduard Komarov <rango.adik@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
